requirements.txt
pillow==11.1.0
torch==2.7.1
torchvision==0.22.1
opencv-python==4.11.0.86

